requirements.txt
amqp==2.5.1
anyjson==0.3.3
argh==0.26.2
arrow==0.15.2
astroid==2.3.2
billiard==3.6.1.0
boto==2.49.0
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Django==2.2.6
django-appconf==1.0.3
django-crispy-forms==1.8.0
django-dramatiq==0.8.0
django-ses==0.8.12
django-timezone-field==3.1
dramatiq==1.8.1
future==0.18.2
gevent==1.4.0
greenlet==0.4.15
html5lib==1.0.1
idna==2.8
importlib-metadata==0.23
isort==4.3.21
kombu==4.6.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==7.2.0
pathtools==0.1.2
pika==1.1.0
Pillow==6.2.1
pkg-resources==0.0.0
prometheus-client==0.7.1
psycopg2==2.8.4
pycparser==2.19
pylint==2.4.3
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
rabbitmq==0.2.0
redis==3.3.11
reportlab==3.5.42
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
stripe==2.45.0
typed-ast==1.4.0
urllib3==1.25.7
vine==1.3.0
watchdog==0.8.3
watchdog-gevent==0.1.0
webencodings==0.5.1
wrapt==1.11.2
xhtml2pdf==0.2.4
zipp==0.6.0
